Update polling_monitor.py for detecting change in files inside folders.
Updates polling_monitor.py so that it is able to recognize if there's change in the files inside the monitored folders. The goal is to improve compatibility with debrid services when adding non cached torrents, in these cases the files are being downloaded and until the download completes, they appear in the directory, so what it does now is to compare the list of files inside each directory.